requirements.txt
django
django_rest_framework
django-cors-headers
djangorestframework-simplejwt
drf-yasg
mysqlclient
django-twilio
otp_twilio